ILEX-13 ui changes
